done tabView + 2 default views
