+++ AcloConcurrency/AcloConcurrency/Collections/Views/SearchView.swift
@@ -1,0 +1,22 @@
+//
+//  SearchView.swift
+//  AlcoConcurrency
+//
+//  Created by DariaMikots on 14.06.2022.
+//
+
+import SwiftUI
+
+struct SearchView: View {
+    var body: some View {
+        NavigationView{
+            Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+        }
+    }
+}
+
+struct SearchView_Previews: PreviewProvider {
+    static var previews: some View {
+        SearchView()
+    }
+}
