+++ AcloConcurrency/AcloConcurrency/Collections/Views/SharingView.swift
@@ -1,0 +1,22 @@
+//
+//  SharingView.swift
+//  AlcoConcurrency
+//
+//  Created by DariaMikots on 14.06.2022.
+//
+
+import SwiftUI
+
+struct SharingView: View {
+    var body: some View {
+        NavigationView{
+            Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+        }
+    }
+}
+
+struct SharingView_Previews: PreviewProvider {
+    static var previews: some View {
+        SharingView()
+    }
+}
